rm old agg
